pyproject.toml
[build-system]
requires = ["setuptools==80.10.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ami-dataops"
version = "0.2.0"
description = "Data operations toolkit — backup, sync, provisioning, monitoring, and maintenance for data services"
requires-python = "==3.11.*"
dependencies = [
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "loguru==0.7.3",
    "pyyaml==6.0.3",
    "httpx==0.28.1",
    "setuptools==80.10.2",
    "python-dotenv==1.2.2",
    "google-auth==2.49.1",
    "google-auth-oauthlib==1.3.0",
    "google-api-python-client==2.193.0",
    "zstandard==0.25.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "ami-ci[dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ami.dataops*"]
namespaces = true

[tool.uv]
package = true

[tool.uv.sources]
ami-ci = { path = "../AMI-CI", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
norecursedirs = ["tmp", "build", "dist"]
